refactor(use-resize-observer): convert to typescript
