+++ src/hooks/__internal__/useResizeObserver/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./useResizeObserver";
